changed conversion of a string(stoi -> stod).

+++ Assignment2b/main.cpp
@@ -43,7 +43,7 @@ int main()
 		                                temp += str[i];
 		                                i++;
 		                        }
-		                        *start = stoi(temp);
+		                        *start = stod(temp);
 		                        start++;
 		        }
 		}
